+++ modules/@apostrophecms/image/ui/apos/components/AposMediaManager.vue
@@ -1,9 +1,3 @@
-<!--
-  AposMediaManager will be in charge of all media-related state logic
-  this includes doing the selecting and deselecting of items, deciding the editor/selection view,
-  emitting batch action events, etc. All sub views will recieve `media` as a prop
--->
-
 <template>
   <AposModal
     :modal="modal"
@@ -13,14 +7,20 @@
     @show-modal="modal.showModal = true"
     @esc="confirmAndCancel"
   >
-    <template v-if="relationshipField" #secondaryControls>
+    <template
+      v-if="relationshipField"
+      #secondaryControls
+    >
       <AposButton
         type="default"
         label="apostrophe:cancel"
         @click="confirmAndCancel"
       />
     </template>
-    <template v-else #secondaryControls>
+    <template
+      v-else
+      #secondaryControls
+    >
       <AposButton
         type="default"
         label="apostrophe:exit"
@@ -52,7 +52,10 @@
     </template>
     <template #main>
       <AposLoadingBlock v-if="isFirstLoading" />
-      <AposModalBody v-else ref="modalBody">
+      <AposModalBody
+        v-else
+        ref="modalBody"
+      >
         <template #bodyHeader>
           <AposDocsManagerToolbar
             :selected-state="selectAllState"
@@ -541,7 +544,10 @@ export default {
     select(id) {
       if (this.checked.includes(id)) {
         this.updateEditing(id);
-      } else if (this.relationshipField && (this.relationshipField.max > 1 || !this.relationshipField.max)) {
+      } else if (
+        this.relationshipField &&
+        (this.relationshipField.max > 1 || !this.relationshipField.max)
+      ) {
         this.selectAnother(id);
       } else {
         this.checked = [ id ];
@@ -669,6 +675,7 @@ export default {
             await this.loadWhenIntersecting();
             await this.$nextTick();
             await this.loadUntilScroll();
+          // eslint-disable-next-line no-console
           }).catch(console.error);
         }
       }
@@ -732,7 +739,8 @@ export default {
               ...requestOptions,
               _ids: this.checked,
               messages,
-              type: this.checked.length === 1 ? this.moduleLabels.singular
+              type: this.checked.length === 1
+                ? this.moduleLabels.singular
                 : this.moduleLabels.plural
             }
           });
@@ -741,7 +749,6 @@ export default {
             interpolate: { operation: this.$t(label) },
             type: 'danger'
           });
-          console.error(error);
         }
       }
     }