+++ modules/@apostrophecms/page/ui/apos/logic/AposPagesManager.js
@@ -119,15 +119,18 @@ export default {
       const checkLen = this.checked.length;
       const rowLen = this.items.length;
 
-      return checkLen > 0 && checkLen !== rowLen ? {
-        value: 'checked',
-        indeterminate: true
-      } : {
-        value: 'checked'
-      };
+      return checkLen > 0 && checkLen !== rowLen
+        ? {
+          value: 'checked',
+          indeterminate: true
+        }
+        : {
+          value: 'checked'
+        };
     },
     canCreate() {
-      const page = this.items.find(page => page.aposDocId === this.moduleOptions.page.aposDocId);
+      const page = this.items
+        .find(page => page.aposDocId === this.moduleOptions.page.aposDocId);
       if (page) {
         return page._create;
       }
@@ -258,15 +261,15 @@ export default {
         return (page.aposDocId === window.apos.page.page?.aposDocId) && page.archived;
       })) {
         // With the current page gone, we need to move to safe ground
-        location.assign(`${window.apos.prefix}/`);
+        location.assign(`${apos.prefix}/`);
       }
     },
-    // Recursively update the tree and the flat list with the changes returned by the server.
-    // The `changes` array contains a list of documents with only the fields
-    // that have changed. Update both draft and published documents. The current
-    // document changes comes only for draft and contains the new depth, updatedAt,
-    // and order. We request a fresh published document for the current document
-    // to update it.
+    // Recursively update the tree and the flat list with the changes returned
+    // by the server. The `changes` array contains a list of documents with
+    // only the fields that have changed. Update both draft and published documents.
+    // The current document changes comes only for draft and contains the new depth,
+    // updatedAt, and order. We request a fresh published document for the current
+    // document to update it.
     // The rest of the changes are only the new order values for both
     // draft and published documents.
     async updateTree(updated) {
@@ -508,7 +511,8 @@ export default {
               ...requestOptions,
               _ids: this.checked,
               messages,
-              type: this.checked.length === 1 ? this.moduleLabels.singular
+              type: this.checked.length === 1
+                ? this.moduleLabels.singular
                 : this.moduleLabels.plural
             }
           });
@@ -522,6 +526,7 @@ export default {
             interpolate: { operation: label },
             type: 'danger'
           });
+          // eslint-disable-next-line no-console
           console.error(error);
         }
       }