+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Allow scrolling while dragging a page in the Page Manager.
 * Change user's email field type to `email`.
 * Improve media manager experience after uploading images. No additional server requests are made, no broken UI on error.
+* Change reset password form button label to `Reset Password`.
 
 ### Fixes